+++ src/main/java/org/lostfan/ktv/domain/Material.java
@@ -1,12 +1,14 @@
 package org.lostfan.ktv.domain;
 
+import java.math.BigDecimal;
+
 public class Material extends DefaultEntity {
 
     private Integer id;
 
     private String name;
 
-    private Integer price;
+    private BigDecimal price;
 
     private String unit;
 
@@ -18,11 +20,11 @@ public class Material extends DefaultEntity {
         this.name = name;
     }
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
