+++ src/main/java/org/lostfan/ktv/model/entity/EquipmentEntityModel.java
@@ -24,7 +24,7 @@ public class EquipmentEntityModel extends BaseEntityModel<Equipment> {
         this.fields = new ArrayList<>();
         this.fields.add(new EntityField("equipment.id", EntityFieldTypes.Integer, Equipment::getId, Equipment::setId, false));
         this.fields.add(new EntityField("equipment.name", EntityFieldTypes.String, Equipment::getName, Equipment::setName));
-        this.fields.add(new EntityField("equipment.price", EntityFieldTypes.Integer, Equipment::getPrice, Equipment::setPrice));
+        this.fields.add(new EntityField("equipment.price", EntityFieldTypes.Double, Equipment::getPrice, Equipment::setPrice));
     }
 
     @Override
