+++ src/main/java/org/lostfan/ktv/model/entity/MaterialEntityModel.java
@@ -23,7 +23,7 @@ public class MaterialEntityModel extends BaseEntityModel<Material> {
         this.fields = new ArrayList<>();
         this.fields.add(new EntityField("material.id", EntityFieldTypes.Integer, Material::getId, Material::setId, false));
         this.fields.add(new EntityField("material.name", EntityFieldTypes.String, Material::getName, Material::setName));
-        this.fields.add(new EntityField("material.price", EntityFieldTypes.Integer, Material::getPrice, Material::setPrice));
+        this.fields.add(new EntityField("material.price", EntityFieldTypes.Double, Material::getPrice, Material::setPrice));
         this.fields.add(new EntityField("material.unit", EntityFieldTypes.String, Material::getUnit, Material::setUnit));
     }
