+++ src/main/java/org/lostfan/ktv/model/entity/TariffEntityModel.java
@@ -1,12 +1,12 @@
 package org.lostfan.ktv.model.entity;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Function;
 
 import org.lostfan.ktv.dao.DAOFactory;
-import org.lostfan.ktv.dao.EntityDAO;
 import org.lostfan.ktv.dao.TariffDAO;
 import org.lostfan.ktv.domain.Tariff;
 import org.lostfan.ktv.domain.TariffPrice;
@@ -41,9 +41,9 @@ public class TariffEntityModel extends BaseEntityModel<Tariff> {
         this.fields.add(new EntityField("tariff.name", EntityFieldTypes.String, Tariff::getName, Tariff::setName));
         this.fields.add(new EntityField("tariff.digital", EntityFieldTypes.Boolean, Tariff::isDigital, Tariff::setDigital));
         this.fields.add(new EntityField("tariff.channels", EntityFieldTypes.String, Tariff::getChannels, Tariff::setChannels));
-        this.fields.add(new EntityField("tariff.currentPrice", EntityFieldTypes.Integer, new Function<Tariff, Integer>() {
+        this.fields.add(new EntityField("tariff.currentPrice", EntityFieldTypes.Double, new Function<Tariff, BigDecimal>() {
             @Override
-            public Integer apply(Tariff tariff) {
+            public BigDecimal apply(Tariff tariff) {
                 return getDao().getPriceByDate(tariff.getId(), LocalDate.now());
             }
         }, (e1,e2) -> {}, false));
