+++ src/main/java/org/lostfan/ktv/model/searcher/MaterialSearcherModel.java
@@ -17,7 +17,7 @@ public class MaterialSearcherModel extends EntitySearcherModel<Material> {
         this.fields = new ArrayList<>();
         this.fields.add(new EntityField("material.id", EntityFieldTypes.Integer, Material::getId, Material::setId, false));
         this.fields.add(new EntityField("material.name", EntityFieldTypes.String, Material::getName, Material::setName));
-        this.fields.add(new EntityField("material.price", EntityFieldTypes.Integer, Material::getPrice, Material::setPrice));
+        this.fields.add(new EntityField("material.price", EntityFieldTypes.Double, Material::getPrice, Material::setPrice));
         this.fields.add(new EntityField("material.unit", EntityFieldTypes.String, Material::getUnit, Material::setUnit));
     }
